feat: introduce the interpreter
the old way to compute a parameter before it convert to bytes,
a preconditon function to check whether the data received is valid and
corresponding postprocess to process a parameter after it is extracted
from bytes is register a corresponding function. This refactor introduce
a brand new way to do the same work, but not hardcoded in source file.
We know that fixed format data is easy to descript in xml file, but how
the code data can be descripted in xml file, they can vastly variable!
The answer is positive, just think about our compilers, to them source
code we written also vastly variable; So if we can write a "complier" we
can also realize it. Here where a very simple interperter comes in, whoes
function is to compile the souce code in xml file into byte code in
internal form, later on when we need it, the interpreter can execute
the byte code for us and return information we will use